+++ src/lib.rs
@@ -15,11 +15,15 @@ mod error;
 mod sys;
 
 pub use codec_info::*;
-pub use decode::{DecodedFrame, Decoder, DecoderCaps, DecoderCodec, DecoderConfig, SurfaceFormat};
+pub use decode::{
+    DecodedFrame, Decoder, DecoderCaps, DecoderCodec, DecoderConfig, SurfaceFormat,
+    query_decoder_caps,
+};
 pub use encode::{
     Av1EncoderConfig, Av1Profile, BufferFormat, CodecConfig, EncodeOptions, EncodedFrame, Encoder,
     EncoderCaps, EncoderCodec, EncoderConfig, H264EncoderConfig, H264Profile, HevcEncoderConfig,
     HevcProfile, PictureType, Preset, RateControlMode, ReconfigureParams, TuningInfo,
+    query_encoder_caps,
 };
 pub use error::Error;